Switch offlineimap -> mbsync for pulling mail
offlineimap is kinda slow (it's written in Python), and it often hangs
for me. I'm switching to mbsync; it does the same thing, but it's
written in C (so it's a bit faster), has relatively comprehensive tests,
and can perform a one-way mail sync (which is much less blood-curdling
than offlineimap when you're first setting it up).