Include another TeX bin directory in PATH
This one contains epstopdf, which is used for incorporating EPS graphics
into LaTeX documents.

I occasionally use gnuplot in org files. That can produce PNGs, which are ugly
and pixelated, or it can produce EPS images, which are vector-based and
comparatively lovely. Those EPS files need to be converted into PDF for pdflatex
to handle them, though. Hence this change.