Shell-escape when producing PDFs
This adds the `-shell-escape` flag to the pdflatex command run when exporting an
org document to PDF.

I want to produce PDFs with syntax highlighting in the code. The best way to do
that seems to be with the "minted" package, but that package shells out to
pygments to do the actual work. pdflatex usually disallows shell commands; this
enables that.