Switch from urlview to extract_url.pl
http://www.memoryhole.net/~kyle/extract_url/

When I want to follow a URL in an email in mutt, I've usually used
urlview to parse out the list and provide a curses interface to browse
them. Unfortunately, urlview doesn't provide any context around the
link, so picking out a specific URL in an email with many similar URLs
is difficult.

extract_url.pl is a drop-in replacement for urlview. It provides the
same functionality, but can *also* be configured to provide textual
context instead of only listing raw URLs.