Use evil with org-agendas
I've never been really into the agenda mode of Org; I usually just modify the
files directly. I've been playing around with it more over the last week or so,
though, and the lack of Vim keybindings is drivin' me nuts.

This installs the `evil-org` package and ensures that it'll be used.