Ensure that new frames use the current font size
Since `add-to-list` won't add a new `font-code` to the `default-frame-alist` if
it's already these, this can result in new frames having different font sizes.

This is also kinda sloppy in general.

This commit explicitly resets the `font` entry in that list, and only appends a
new entry if we haven't already set one.

Thanks to @rneatherway for the nice solution! Fixes #50.